requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
crewai==0.186.1
crewai-tools==0.71.0
langchain-community==0.3.29
langchain-google-genai==2.1.12
pandas==2.3.2
openpyxl==3.1.5
pypdf==5.9.0
python-dotenv==1.0.0
requests==2.32.5
gunicorn==21.2.0
tokenizers==0.20.3
openai-whisper==20250625
pyaudio==0.2.14
ffmpeg-python==0.2.0